# spec/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Urigin
  license:
    name: MIT
servers:
  - url: http://localhost:3000/v1
paths:
...
  /users/{userId}/games:
    get:
      summary: Get all game from a user
      operationId: get game from a user
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: Numeric ID of the user to get
      tags:
        - users
      security:
        - bearerAuth: [user, creator, moderator, admin]
      responses:
        '200':
          description: list of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '403':
          description: Forbidden, you don't have enough rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '404':
          description: user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '498':
          description: Your token is invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
    post:
      summary: Add a new game in your library
      operationId: add game in your library
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: Numeric ID of the user to get
      tags:
        - users
      security:
        - bearerAuth: [user, creator, moderator, admin]
      requestBody:
        description: add new game
        required: true
        content:
          application/json:
            schema:
              properties:
                gameId:
                  type: integer
                favorite:
                  type: boolean
                score:
                  type: integer
      responses:
        '200':
          description: creation success
        '400':
          description: error with the body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '403':
          description: Forbidden, you don't have enough rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '404':
          description: user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '498':
          description: Your token is invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
  /users/{userId}/games/{gameId}:
    put:
      summary: update your game favorite or scored
      operationId: update games relation
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: Numeric ID of the user to get
        - in: path
          name: gameId
          schema:
            type: integer
          required: true
          description: Numeric ID of your game
      tags:
        - users
      security:
        - bearerAuth: [user, creator, moderator, admin]
      requestBody:
        description: updated relation
        required: true
        content:
          application/json:
            schema:
              properties:
                favorite:
                  type: boolean
                score:
                  type: integer
      responses:
        '200':
          description: updated success
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '400':
          description: error with the body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '403':
          description: Forbidden, you don't have enough rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '404':
          description: game or user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '498':
          description: Your token is invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
    delete:
      summary: delete your game from your library
      operationId: delete game relation
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: Numeric ID of the user to get
        - in: path
          name: gameId
          schema:
            type: integer
          required: true
          description: Numeric ID of your game
      tags:
        - users
      security:
        - bearerAuth: [user, creator, moderator, admin]
      responses:
        '200':
          description: deleted success
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '403':
          description: Forbidden, you don't have enough rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '404':
          description: relation or user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '498':
          description: Your token is invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
...
  /games:
    get:
      summary: Get all games
      operationId: get games
      tags:
        - games
      responses:
        '200':
          description: list of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
        '404':
          description: game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
    post:
      summary: Add a new game
      operationId: add game
      tags:
        - games
      security:
        - bearerAuth: [creator]
      requestBody:
        description: add new game
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: creation success
        '400':
          description: error with the body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '403':
          description: Forbidden, you don't have enough rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '498':
          description: Your token is invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
  /games/{gameId}:
    put:
      summary: update a game
      operationId: update a game
      tags:
        - games
      parameters:
        - in: path
          name: gameId
          schema:
            type: integer
          required: true
          description: Numeric ID of the game to get
      security:
        - bearerAuth: [creator, admin]
      requestBody:
        description: update game
        required: true
        content:
          application/json:
            schema:
              properties:
                description:
                  type: string
      responses:
        '200':
          description: update success
        '400':
          description: error with the body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '403':
          description: Forbidden, you don't have enough rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '498':
          description: Your token is invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
    delete:
      summary: delete a game
      operationId: delete a game
      tags:
        - games
      parameters:
        - in: path
          name: gameId
          schema:
            type: integer
          required: true
          description: Numeric ID of the game to get
      security:
        - bearerAuth: [creator, admin]
      responses:
        '200':
          description: delete success
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '403':
          description: Forbidden, you don't have enough rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '404':
          description: game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '498':
          description: Your token is invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
  /games/{gameId}/types:
    get:
      summary: Get all types from a game
      operationId: get games types
      tags:
        - games
      parameters:
        - in: path
          name: gameId
          schema:
            type: integer
          required: true
          description: Numeric ID of the game to get
      responses:
        '200':
          description: list of types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Type"
        '404':
          description: game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
    post:
      summary: Add a new type to a game
      operationId: add game type
      parameters:
        - in: path
          name: gameId
          schema:
            type: integer
          required: true
          description: Numeric ID of the game to get
      tags:
        - games
      security:
        - bearerAuth: [creator]
      requestBody:
        description: add new game type
        required: true
        content:
          application/json:
            schema:
              properties:
                typeId:
                  type: integer
      responses:
        '200':
          description: link success
        '400':
          description: error with the body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '403':
          description: Forbidden, you don't have enough rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '498':
          description: Your token is invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
...
  /games/{gameId}/users:
    get:
      summary: Get all users from a game
      operationId: get games users
      tags:
        - games
      parameters:
        - in: path
          name: gameId
          schema:
            type: integer
          required: true
          description: Numeric ID of the game to get
      responses:
        '200':
          description: list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        '404':
          description: game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
  /games/{gameId}/comments:
    get:
      summary: Get all comments from a game
      operationId: get games comments
      tags:
        - games
      parameters:
        - in: path
          name: gameId
          schema:
            type: integer
          required: true
          description: Numeric ID of the game to get
      responses:
        '200':
          description: list of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        '404':
          description: game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
    post:
      summary: add new comments on a game
      operationId: add game comments
      parameters:
        - in: path
          name: gameId
          schema:
            type: integer
          required: true
          description: Numeric ID of the game to get
      tags:
        - games
      security:
        - bearerAuth: [user, creator, admin]
      requestBody:
        description: add new comments
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Comment"
      responses:
        '200':
          description: creation success
        '400':
          description: error with the body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '403':
          description: Forbidden, you don't have enough rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '404':
          description: game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '498':
          description: Your token is invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
  /games/{gameId}/comments/{commentId}:
    put:
      summary: update a comment from a game
      operationId: update a comments
      parameters:
        - in: path
          name: gameId
          schema:
            type: integer
          required: true
          description: Numeric ID of the game to get
        - in: path
          name: commentId
          schema:
            type: integer
          required: true
          description: Numeric ID of the comment to get
      tags:
        - games
      security:
        - bearerAuth: [user, creator, admin]
      requestBody:
        description: update comment
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Comment"
      responses:
        '200':
          description: update success
        '400':
          description: error with the body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '403':
          description: Forbidden, you don't have enough rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '404':
          description: game or comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '498':
          description: Your token is invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
    delete:
      summary: delete a comment from a game
      operationId: delete a comment
      parameters:
        - in: path
          name: gameId
          schema:
            type: integer
          required: true
          description: Numeric ID of the game to get
        - in: path
          name: commentId
          schema:
            type: integer
          required: true
          description: Numeric ID of the comment to get
      tags:
        - games
      security:
        - bearerAuth: [user, creator, admin]
      responses:
        '200':
          description: delete success
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '403':
          description: Forbidden, you don't have enough rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '404':
          description: game or comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
        '498':
          description: Your token is invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UriginError"
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
...
  schemas:
    UriginError:
      required:
        - description
      properties:
        description:
          type: string
    User:
      properties:
        id:
          # Returned by GET, not used in POST/PUT/PATCH
          type: integer
          readOnly: true
        email:
          # Used in POST/PUT/PATCH, not returned by GET
          type: string
          writeOnly: true
        username:
          type: string
        role:
          type: integer
          description: Only if your are admin
        password:
          # Used in POST/PUT/PATCH, not returned by GET
          type: string
          writeOnly: true
...
    Game:
      properties:
        id:
          # Returned by GET, not used in POST/PUT/PATCH
          type: integer
          readOnly: true
        name:
          # Returned by GET, not used in POST/PUT/PATCH
          type: integer
          readOnly: true
        description:
          type: string
          description: Only if you are the creator
        createdAt:
          # Returned by GET, not used in POST/PUT/PATCH
          type: string
          description: creation date of this game
          readOnly: true
    Comment:
      properties:
        id:
          # Returned by GET, not used in POST/PUT/PATCH
          type: integer
          readOnly: true
        content:
          type: string
        createdAt:
          # Returned by GET, not used in POST/PUT/PATCH
          type: string
          description: creation date of this comment
          readOnly: true
        updatedAt:
          # Returned by GET, not used in POST/PUT/PATCH
          type: string
          description: updated date of this comment
          readOnly: true
    Type:
      properties:
        id:
          # Returned by GET, not used in POST/PUT/PATCH
          type: integer
          readOnly: true
        name:
          # Returned by GET, not used in POST/PUT/PATCH
          type: string
          readOnly: true
        description:
          type: string